+++ lib/ethpy/ethpy/hyperdrive/interface/__init__.py
@@ -1,2 +1,2 @@
 """High-level interface for the Hyperdrive market."""
-from .interface import HyperdriveReadInterface
+from .hyperdrive_read_interface import HyperdriveReadInterface
