+++ lib/ethpy/ethpy/hyperdrive/interface/interface_test.py
@@ -14,7 +14,7 @@ from hypertypes.utilities.conversions import (
     pool_info_to_fixedpoint,
 )
 
-from .interface import HyperdriveReadInterface
+from .hyperdrive_read_interface import HyperdriveReadInterface
 
 # we need to use the outer name for fixtures
 # pylint: disable=redefined-outer-name
